apps/backend/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
dj-rest-auth==7.0.1
Django==5.2.1
django-allauth==65.8.1
django-cors-headers==4.7.0
django-filter==25.1
django-stubs==5.2.0
django-stubs-ext==5.2.0
djangorestframework==3.16.0
djangorestframework-stubs==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
mypy==1.15.0
mypy_extensions==1.1.0
mysqlclient==2.2.4 # Added for MariaDB/MySQL
pillow==11.2.1
python-dotenv==1.0.1 # Added for environment variable management
PyJWT==2.9.0
requests==2.32.3
ruff==0.11.11
sqlparse==0.5.3
types-PyYAML==6.0.12.20250516
types-requests==2.32.0.20250515
typing_extensions==4.13.2
urllib3==2.4.0
